Module-6/Demo-2/MCP-Demo/requirements.txt
mcp
fastmcp
requests
google-genai
python-dotenv